@@ -29,7 +29,7 @@ namespace NebulousIndustries.AdventOfCode.Year2020
             long result = 1;
             IEnumerable<TrainTicket> validTickets = this.GetInput().Where(t => t.IsValid);
             List<TrainTicketRule> unassignedRules = new(TrainTicket.Rules);
-            List<TrainTicketRule> assignedRules = new();
+            List<TrainTicketRule> assignedRules = [];
             do
             {
                 (TrainTicketRule assignedRule, int assignedIndex) = FindRulePosition(validTickets, unassignedRules, TrainTicket.Rules.Count);
@@ -101,7 +101,7 @@ namespace NebulousIndustries.AdventOfCode.Year2020
             {
                 return true;
             }
-            else if (input.IndexOf(':') != -1)
+            else if (input.Contains(':'))
             {
                 // Parse a rule
                 string[] ranges = input[(input.IndexOf(": ") + 2)..].Split(" or ");
