@@ -23,40 +23,29 @@ namespace NebulousIndustries.AdventOfCode.Year2021
         {
             IDictionary<string, InsertionRule> rules = this.GetInput().ToDictionary(k => k.AdjacentPair);
 
-            Dictionary<char, long> counts = new();
+            Dictionary<char, long> counts = [];
             foreach (char c in InsertionRule.Template)
             {
-                if (!counts.ContainsKey(c))
-                {
-                    counts[c] = 0;
-                }
-
+                counts.TryAdd(c, 0);
                 counts[c]++;
             }
 
             foreach (char c in rules.Values.Select(r => r.ToInsert))
             {
-                if (!counts.ContainsKey(c))
-                {
-                    counts[c] = 0;
-                }
+                counts.TryAdd(c, 0);
             }
 
-            Dictionary<string, long> pairs = new();
+            Dictionary<string, long> pairs = [];
             for (int i = 1; i < InsertionRule.Template.Length; i++)
             {
                 string pair = InsertionRule.Template.Substring(i - 1, 2);
-                if (!pairs.ContainsKey(pair))
-                {
-                    pairs[pair] = 0;
-                }
-
+                pairs.TryAdd(pair, 0);
                 pairs[pair]++;
             }
 
             for (int s = 0; s < steps; s++)
             {
-                Dictionary<string, long> newPairs = new();
+                Dictionary<string, long> newPairs = [];
 
                 foreach (KeyValuePair<string, long> pair in pairs)
                 {
@@ -64,16 +53,10 @@ namespace NebulousIndustries.AdventOfCode.Year2021
                     counts[rule.ToInsert] += pair.Value;
 
                     string left = string.Concat(pair.Key[0], rule.ToInsert);
-                    if (!newPairs.ContainsKey(left))
-                    {
-                        newPairs[left] = 0;
-                    }
+                    newPairs.TryAdd(left, 0);
 
                     string right = string.Concat(rule.ToInsert, pair.Key[1]);
-                    if (!newPairs.ContainsKey(right))
-                    {
-                        newPairs[right] = 0;
-                    }
+                    newPairs.TryAdd(right, 0);
 
                     newPairs[left] += pair.Value;
                     newPairs[right] += pair.Value;
